+++ src/display/col.rs
@@ -99,8 +99,8 @@ pub type Cols = [Col;COLS_COUNT];
 pub static DEFAULT_COLS: Cols = [
     Col::Mark,
     Col::Git,
-    Col::Permission,
     Col::Date,
+    Col::Permission,
     Col::Size,
     Col::Count,
     Col::Branch,