+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### next
 - when sizes are displayed (eg on `br -s`) show size of root line and root filesystem info
+- modified size cache management makes some size computations faster
 
 <a name="v1.0.7"></a>
 ### v1.0.7 - 2020-11-27